+++ packages/string/src/index.ts
@@ -1,1 +1,2 @@
+export { camelCase } from './camel-case';
 export { capitalize } from './capitalize';